+++ src/main/java/nordmods/uselessreptile/datagen/data/mod/UREquipmentProvider.java
@@ -63,7 +63,7 @@ public class UREquipmentProvider implements DataProvider {
     }
 
     protected void addEntry(EntityType<? extends URDragonEntity> type, Item item, Identifier texture, Identifier model, boolean translucent) {
-        DragonEquipment.Equipment equipmentModelData = new DragonEquipment.Equipment(Registries.ITEM.getId(item), new ModelData(texture, model, Optional.empty(), false, translucent));
+        DragonEquipment.Equipment equipmentModelData = new DragonEquipment.Equipment(Registries.ITEM.getId(item), new ModelData(texture, model, Optional.empty(), true, translucent));
         Identifier id = EntityType.getId(type);
         if (holder.containsKey(id)) holder.get(id).equipment().add(equipmentModelData);
         else holder.put(id, new DragonEquipment(Optional.empty(), new ArrayList<>(Collections.singleton(equipmentModelData))));